App updates and fix for new bug where submission status changes not accepted

+++ Website/OCM.MVC/Controllers/LoginProviderController.cs
@@ -87,7 +87,7 @@ namespace OCM.MVC.Controllers
                     }
 
                     string screenName;
-                    int userId;
+                    long userId;
                     if (TwitterConsumer.TryFinishSignInWithTwitter(out screenName, out userId))
                     {
                         string newSessionToken = Guid.NewGuid().ToString();
